Move widget to addon
