+++ app/components/flexberry-layers-attributes-panel.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis/components/flexberry-layers-attributes-panel';
