+++ tests/dummy/app/controllers/map.js
@@ -93,84 +93,11 @@ export default EditMapController.extend(
       return result;
     }),
 
-    /**
-     * bottompanel items
-     */
-    bottompanelItems: Ember.computed('bottompanel.[]', 'bottompanel.@each.active', 'i18n', function () {
-      let result = Ember.A([{
-          selector: 'attr',
-          captionPath: 'forms.map.bottompanelbuttontooltip',
-          iconClass: 'browser icon'
-        }]);
+    editedLayers: Ember.A(),
 
-      return result;
-    }),
+    editedLayersSelectedTabIndex: 0,
 
-    _editedLayers: Ember.A(),
-
-    _editedLayersActiveTabIndex: 0,
-
-    _editedLayersTabs: Ember.computed('_editedLayers.[]', function() {
-      let editedLayers = this.get('_editedLayers');
-      if (Ember.isPresent(editedLayers)) {
-        return editedLayers.map((item) => {
-          let name = Ember.get(item, 'name');
-          let header = {};
-          let featureLink = {};
-          let propertyLink = {};
-          let properties = Ember.A();
-          let leafletObject = Ember.get(item, 'leafletObject');
-
-          leafletObject.eachLayer((l) => {
-            let props = Ember.get(l, 'feature.properties');
-            let propId = Ember.guidFor(props);
-            if (Ember.isNone(l.feature.leafletLayer)) {
-              Ember.set(l.feature, 'leafletLayer', l);
-            }
-
-            // the hash containing guid of properties object and link to feature layer
-            featureLink[propId] = l;
-
-            // the hash containing guid of properties object and link to that object
-            propertyLink[propId] = props;
-
-            // collect the object keys for the header object
-            Object.keys(props).forEach((p) => {
-              if (!header.hasOwnProperty(p)) {
-                header[p] = p;
-              }
-            });
-            properties.pushObject(props);
-          });
-          let tabModel = Ember.Object.extend({
-            _top: 5,
-            _skip: 0,
-            _selectedRows: {},
-            _selectedRowsCount: Ember.computed('_selectedRows', function() {
-              let selectedRows = Ember.get(this, '_selectedRows');
-              return Object.keys(selectedRows).filter((item) => Ember.get(selectedRows, item)).length;
-            }),
-            selectAll: false,
-            propertiesToShow: Ember.computed('properties.[]', '_top', '_skip', function() {
-              let properties = this.get('properties');
-              let top = this.get('_top');
-              let skip = this.get('_skip');
-              return properties.slice(skip, skip + top);
-            })
-          });
-          return tabModel.create({
-            name: name,
-            leafletObject: leafletObject,
-            featureLink: featureLink,
-            propertyLink: propertyLink,
-            header: header,
-            properties: properties
-          });
-        });
-      }
-    }),
-
-    _featureTabsOffset: 0,
+    editedLayersPanelFolded: true,
 
     availableCRS: Ember.computed('i18n.locale', function () {
       let availableModes = Ember.A();
@@ -192,20 +119,6 @@ export default EditMapController.extend(
 
       return availableModes;
     }),
-
-    // code from layer-result-list
-    /**
-      Set selected feature and add its layer to serviceLayer on map.
-      @method _selectFeature
-      @param {Object} feature Describes inner FeatureResultItem's feature object or array of it.
-      @private
-    */
-    _selectFeature(feature) {
-      let serviceLayer = this.get('serviceLayer');
-      if (!Ember.isNone(feature)) {
-        serviceLayer.addLayer(feature.leafletLayer.setStyle({ color:'salmon', weight:2, opacity:1, fillOpacity:0.3 }));
-      }
-    },
 
     actions: {
       toggleSidebar(e) {
@@ -243,181 +156,6 @@ export default EditMapController.extend(
             }, 500);
           }
         }
-      },
-
-      toggleBottompanel() {
-        this.set('bottompanelOpened', !this.get('bottompanelOpened'));
-      },
-
-      activeTab(index) {
-        if (index === this.get('_editedLayersActiveTabIndex')) {
-          this.send('toggleBottompanel');
-        } else {
-          this.set('_editedLayersActiveTabIndex', index);
-          if (!this.get('bottompanelOpened')) {
-            this.send('toggleBottompanel');
-          }
-        }
-      },
-
-      closeTab(index) {
-        let editedLayers = this.get('_editedLayers');
-
-        if (this.get('_editedLayersActiveTabIndex') < index) {
-          editedLayers.removeAt(index);
-        } else {
-          editedLayers.removeAt(index);
-          if (this.get('_editedLayersActiveTabIndex') - 1 >= 0) {
-            this.set('_editedLayersActiveTabIndex', this.get('_editedLayersActiveTabIndex') - 1);
-          }
-        }
-      },
-
-      getAttributes(object) {
-        let editedLayers = this.get('_editedLayers') || Ember.A();
-        let index = editedLayers.findIndex((item) => Ember.isEqual(item.name, object.name));
-        if (index >= 0) {
-          this.send('activeTab', index);
-        } else {
-          editedLayers.addObject(object);
-          this.set('_editedLayers', editedLayers);
-          this.set('_editedLayersActiveTabIndex', editedLayers.length - 1);
-        }
-
-        if (!this.get('bottompanelOpened')) {
-          this.send('toggleBottompanel');
-        }
-      },
-
-      onFeatureRowSelect(tabModel, rowId, options) {
-        let selectedRows = Ember.get(tabModel, '_selectedRows');
-        Ember.set(selectedRows, rowId, options.checked);
-        Ember.set(tabModel, '_selectedRows', selectedRows);
-        tabModel.notifyPropertyChange('_selectedRows');
-      },
-
-      onFindFeatureClick(tabModel) {
-        let selectedRows = Ember.get(tabModel, '_selectedRows');
-        let selectedFeatures = Object.keys(selectedRows).filter((item) => Ember.get(selectedRows, item))
-          .map((key) => {
-            return tabModel.featureLink[key].feature;
-          });
-        this.send('zoomTo', selectedFeatures);
-      },
-
-      onFeatureSelectClick(tabModel) {
-        let selectAll = Ember.get(tabModel, 'selectAll');
-        if (!selectAll) {
-          Ember.set(tabModel, '_selectedRows', {});
-        } else {
-          let selectedRows = Object.assign(...Object.keys(Ember.get(tabModel, 'propertyLink')).map(k => ({ [k]: true })));
-          Ember.set(tabModel, '_selectedRows', selectedRows);
-        }
-
-        tabModel.notifyPropertyChange('_selectedRows');
-      },
-
-      onFeatureGetData(tabModel, options) {
-        Ember.set(tabModel, '_skip', options.skip);
-        Ember.set(tabModel, '_top', options.top);
-      },
-
-      onFeatureTabMove(left) {
-        // move tabs bar on the left or on the right
-        let offset = this.get('_featureTabsOffset');
-        if (left) {
-          if (offset > 0) {
-            offset -= Math.min(25, offset);
-          }
-        } else {
-          let panelWidth = Ember.$('.feature-tab-nav-panel-tabs').innerWidth();
-          let itemsWidth = 0;
-          Ember.$('.feature-tab-nav-panel-tabs').children().each((index, item) => {
-            itemsWidth += Ember.$(item).outerWidth();
-          });
-          offset += Math.min(25, (panelWidth >= itemsWidth ? 0 : (offset >= itemsWidth - panelWidth + 25 ? 0 : itemsWidth - panelWidth + 25 - offset)));
-        }
-
-        this.set('_featureTabsOffset', offset);
-        Ember.$('.feature-tab-nav-panel-tabs').css('left', `-${offset}px`);
-      },
-
-      ///////////// code from layer-result-list
-      /**
-        Handles inner FeatureResultItem's bubbled 'selectFeature' action.
-        Invokes component's {{#crossLink "FeatureResultItem/sendingActions.selectFeature:method"}}'selectFeature'{{/crossLink}} action.
-
-        @method actions.selectFeature
-        @param {Object} feature Describes inner FeatureResultItem's feature object or array of it.
-      */
-      selectFeature(feature) {
-        let selectedFeature = this.get('_selectedFeature');
-        let serviceLayer = this.get('serviceLayer');
-
-        if (selectedFeature !== feature) {
-          if (!Ember.isNone(serviceLayer)) {
-            serviceLayer.clearLayers();
-
-            if (Ember.isArray(feature)) {
-              feature.forEach((item) => this._selectFeature(item));
-            } else {
-              this._selectFeature(feature);
-            }
-          }
-
-          this.set('_selectedFeature', feature);
-        }
-
-        // Send action despite of the fact feature changed or not. User can disable layer anytime.
-        // this.sendAction('featureSelected', feature);
-      },
-
-      /**
-        Select passed feature and zoom map to its layer bounds
-        @method actions.zoomTo
-        @param {Object} feature Describes inner FeatureResultItem's feature object or array of it.
-      */
-      zoomTo(feature) {
-        let serviceLayer = this.get('serviceLayer');
-        if (!serviceLayer) {
-          let leafletMap = this.get('leafletMap');
-          serviceLayer = L.featureGroup().addTo(leafletMap);
-          this.set('serviceLayer', serviceLayer);
-        } else {
-          serviceLayer.clearLayers();
-        }
-
-        this.send('selectFeature', feature);
-
-        let bounds;
-        if (typeof (serviceLayer.getBounds) === 'function') {
-          bounds = serviceLayer.getBounds();
-        } else {
-          let featureGroup = L.featureGroup(serviceLayer.getLayers());
-          if (featureGroup) {
-            bounds = featureGroup.getBounds();
-          }
-        }
-
-        if (!Ember.isBlank(bounds)) {
-          // 'bound.pad(1)' bounds with zoom decreased by 1 point (padding).
-          //  That allows to make map's bounds slightly larger than serviceLayer's bounds to make better UI.
-          this.get('leafletMap').fitBounds(bounds.pad(1));
-        }
-      },
-
-      ///////////// code from layer-result-list
-
-      onDeleteFeatureClick(tabModel) {
-        let selectedRows = Ember.get(tabModel, '_selectedRows');
-        let selectedFeatureKeys = Object.keys(selectedRows).filter((item) => Ember.get(selectedRows, item));
-        selectedFeatureKeys.forEach((key) => {
-          tabModel.leafletObject.removeLayer(tabModel.featureLink[key]);
-          delete selectedRows[key];
-          tabModel.properties.removeObject(tabModel.propertyLink[key]);
-        });
-        Ember.set(tabModel, '_selectedRows', selectedRows);
-        tabModel.notifyPropertyChange('_selectedRows');
       },
 
       querySearch(queryString) {
@@ -489,6 +227,22 @@ export default EditMapController.extend(
             changed: false,
             tabName: 'identify'
           });
+        }
+      },
+
+      getAttributes(object) {
+        let editedLayers = this.get('editedLayers') || Ember.A();
+        let index = editedLayers.findIndex((item) => Ember.isEqual(item.name, object.name));
+        if (index >= 0) {
+          this.set('editedLayersSelectedTabIndex', index);
+        } else {
+          editedLayers.addObject(object);
+          this.set('editedLayers', editedLayers);
+          this.set('editedLayersSelectedTabIndex', editedLayers.length - 1);
+        }
+
+        if (this.get('editedLayersPanelFolded')) {
+          this.set('editedLayersPanelFolded', false);
         }
       }
     }
