+++ tests/dummy/app/templates/map.hbs
@@ -131,82 +131,14 @@
     </div>
   </div>
 
-  <div class={{concat "bottompanel-wrapper" (if (eq _editedLayers.length 0) " hidden" "")}}>
-    <div class={{concat "ui very wide bottom bottompanel pushable tabbar" (if bottompanelOpened "" " invisibly")}}>
-      <button
-        class={{concat "feature-tab-nav-button feature-tab-nav-button-left" (if (eq _featureTabsOffset 0) " disabled" "")}}
-        {{action "onFeatureTabMove" true}}>
-        <i class="left arrow icon"></i>
-      </button>
-      <div class="feature-tab-nav-panel">
-        <div class="ui top attached tabular menu feature-tab-nav-panel-tabs">
-          {{#each _editedLayersTabs key="name" as |tabModel index|}}
-            <a
-              class={{concat "bottompanel tab item" (if (eq _editedLayersActiveTabIndex index) " active" "")}}
-              data-tab="{{tabModel.name}}"
-              {{action "activeTab" index}}>
-              {{tabModel.name}}
-              <i class="icon close" {{action "closeTab" index bubbles=false}}></i>
-            </a>
-          {{/each}}
-        </div>
-      </div>
-      <button
-        class="feature-tab-nav-button feature-tab-nav-button-right"
-        {{action "onFeatureTabMove" false}}>
-        <i class="right arrow icon"></i>
-      </button>
-
-      {{#each _editedLayersTabs key="name" as |tabModel index|}}
-        <div data-tab="{{tabModel.name}}" class={{concat "ui bottom attached tab segment feature-tab-data-panel" (if (eq _editedLayersActiveTabIndex index) " active" "")}}>
-          {{#flexberry-table
-            allowEdit="true"
-            perPageAvailable=false
-            model=tabModel.propertiesToShow
-            header=tabModel.header
-            recordsTotalCount=tabModel.properties.length
-            getData=(action "onFeatureGetData" tabModel)
-            additionalColumnsCount=1
-          }}
-            {{#block-slot "head-block"}}
-              <div class="feature-tab-head-panel">
-                {{flexberry-checkbox
-                  label=(if tabModel.selectAll "Clear" "Select all")
-                  value=tabModel.selectAll
-                  onChange=(action "onFeatureSelectClick" tabModel)
-                  class=(if (eq tabModel.properties.length 0) "disabled")
-                }}
-                <div style="display:inline-block;">
-                  {{flexberry-button
-                    tooltip="Find feature on the map"
-                    click=(action "onFindFeatureClick" tabModel)
-                    iconClass="marker icon"
-                    readonly=(eq tabModel._selectedRowsCount 0)
-                  }}
-                  {{flexberry-button
-                    tooltip="Delete feature"
-                    click=(action "onDeleteFeatureClick" tabModel)
-                    iconClass="trash icon"
-                    readonly=(eq tabModel._selectedRowsCount 0)
-                  }}
-                </div>
-              </div>
-            {{/block-slot}}
-            {{#block-slot "column-header-head-0"}}
-              <th class="narrow">Select</th>
-            {{/block-slot}}
-            {{#block-slot "column-cell-head-0" as |row|}}
-              <td class="narrow">
-                {{flexberry-checkbox
-                  value=(get tabModel._selectedRows (ember-guid row))
-                  onChange=(action "onFeatureRowSelect" tabModel (ember-guid row))
-                }}
-              </td>
-            {{/block-slot}}
-          {{/flexberry-table}}
-        </div>
-      {{/each}}
-    </div>
+  <div class={{concat "bottompanel-wrapper" (if (eq editedLayers.length 0) " hidden" "")}}>
+    {{flexberry-layers-attributes-panel
+      items=editedLayers
+      serviceLayer=serviceLayer
+      leafletMap=leafletMap
+      selectedTabIndex=editedLayersSelectedTabIndex
+      folded=editedLayersPanelFolded
+    }}
   </div>
 
   <div class="row">
